requirements.txt
streamlit
pandas
numpy
pydicom
Pillow
scikit-learn
tensorflow==2.12.0